// components/types.ts
export interface Shape {
  id: string;
  name: string;
  icon: string;
}

export interface Color {
  id: string;
  name: string;
  value: string;
}

export interface Texture {
  id: string;
  name: string;
  value: string;
}

export interface Pattern {
  id: string;
  name: string;
  value: string;
}